+++ src/js/metabolism/map.js
@@ -18,11 +18,11 @@
  *
  */
 
-(function ($) {
+(function($){
 
     googlemap_init = false;
 
-    var Map = function ($map, template, config, callback) {
+    var Map = function($map, template, config, callback){
 
         var self = this;
 
@@ -30,43 +30,48 @@
 
         self.config = {
             marker : {
-                url      : app.asset + '/media/icon/marker@2x.png',
-                hover_url: app.asset + '/media/icon/marker--hover@2x.png',
-                width    : 45,
-                height   : 65,
-                label    : true
+                url       : app.asset+'/media/icon/marker@2x.png',
+                hover_url : app.asset+'/media/icon/marker--hover@2x.png',
+                width     : 45,
+                height    : 65,
+                label_origin : 'center',
+                label: false
             },
-            overlay: {
-                html: '',
-                x   : 0,
-                y   : 0
+            overlay : {
+                html       : '',
+                x          : 0,
+                y          : 0
             },
-            map    : {
-                center           : [
-                    25,
-                    0
-                ],
-                zoom             : 3,
-                maxZoom          : 21,
-                minZoom          : 3,
-                mapTypeId        : 'ROADMAP',
-                mapTypeControl   : true,
-                navigationControl: false,
-                scrollwheel      : true,
-                zoomControl      : true,
-                panControl       : false,
-                streetViewControl: true,
-                styles           : []
-            }
+            map : {
+                center            :[25,0],
+                zoom              : 3,
+                maxZoom           : 21,
+                minZoom           : 3,
+                mapTypeId         : 'ROADMAP',
+                mapTypeControl    : true,
+                navigationControl : false,
+                scrollwheel       : true,
+                zoomControl       : true,
+                panControl        : false,
+                streetViewControl : true,
+                styles            : []
+            },
+            cluster : {
+                classname: 'map-cluster',
+                size: 200,
+                x: 0,
+                y: 0
+            },
+            click_to_use : false
         };
 
         self.context = {
-            cluster : false,
-            markers : [],
-            marker  : false,
-            overlays: [],
-            map     : false,
-            gmap    : false
+            cluster  : false,
+            markers  : [],
+            marker   : false,
+            selected_marker: false,
+            map      : false,
+            gmap     : false
         };
 
 
@@ -76,8 +81,10 @@
         /**
          *
          */
-        self.__construct = function ($map, config, callback) {
-            if (typeof $.gmap3 == 'undefined') {
+        self.__construct =  function( $map, config, callback )
+        {
+            if( typeof $.gmap3 == 'undefined' )
+            {
                 console.warn('gmap3 is required');
                 return;
             }
@@ -86,124 +93,145 @@
 
             var google_api_key = 'google_api' in app ? app.google_api : ('google_key' in app ? app.google_key : '');
 
-            if (!googlemap_init) {
-                googlemap_init = function () { $(document).trigger('google.maps.initialized') };
+            if( !googlemap_init )
+            {
+                googlemap_init = function(){ $(document).trigger('google.maps.initialized') };
 
-                if (document.readyState == 'complete') {
-                    $('head')
-                        .append('<script src="https://maps.google.com/maps/api/js?key=' + google_api_key + '&callback=googlemap_init"></script>');
+                if( document.readyState == 'complete' )
+                {
+                    $('head').append('<script src="https://maps.google.com/maps/api/js?key='+google_api_key+'&callback=googlemap_init"></script>');
                 }
-                else {
-                    $(window).on('load', function () {
-                        $('head')
-                            .append('<script src="https://maps.google.com/maps/api/js?key=' + google_api_key + '&callback=googlemap_init"></script>');
+                else
+                {
+                    $(window).on('load', function()
+                    {
+                        $('head').append('<script src="https://maps.google.com/maps/api/js?key='+google_api_key+'&callback=googlemap_init"></script>');
                     });
                 }
             }
 
-            if (typeof google != 'undefined' && 'maps' in google) {
+            if( typeof google != 'undefined' && 'maps' in google ){
 
-                self.init($map, callback);
+                self._init($map, callback);
             }
-            else {
+            else{
 
                 $(document).on('google.maps.initialized', function () {
 
-                    self.init($map, callback);
+                    self._init($map, callback);
                 });
             }
         };
 
 
+
         /**
          *
          */
-        self.init = function ($map, callback) {
-
+        self._init = function( $map, callback )
+        {
             self.config.map.mapTypeId = google.maps.MapTypeId[self.config.map.mapTypeId];
 
-            self.config.marker_hover = {
-                url        : self.config.marker.hover_url,
-                scaledSize : new google.maps.Size(self.config.marker.width, self.config.marker.height),
-                anchor     : new google.maps.Point(self.config.marker.width / 2, self.config.marker.height),
-                labelOrigin: new google.maps.Point(self.config.marker.width / 2, self.config.marker.height * 0.65)
+            self.context.marker = {
+                url         : self.config.marker.url,
+                scaledSize  : new google.maps.Size(self.config.marker.width, self.config.marker.height),
+                anchor      : new google.maps.Point(self.config.marker.width/2, self.config.marker.height),
+                labelOrigin : new google.maps.Point(self.config.marker.width/2, self.config.marker.height*0.65)
             };
 
-            self.config.marker = {
-                url        : self.config.marker.url,
-                scaledSize : new google.maps.Size(self.config.marker.width, self.config.marker.height),
-                anchor     : new google.maps.Point(self.config.marker.width / 2, self.config.marker.height),
-                labelOrigin: new google.maps.Point(self.config.marker.width / 2, self.config.marker.height * 0.65)
-            };
+            if( self.config.marker.label_origin )
+            {
+                if( self.config.marker.label_origin == 'top' )
+                    self.context.marker.labelOrigin = new google.maps.Point(self.config.marker.width/2, -20);
+                else if( self.config.marker.label_origin == 'bottom' )
+                    self.context.marker.labelOrigin = new google.maps.Point(self.config.marker.width/2, self.config.marker.height+10);
+            }
 
             self.context.$map = $map;
 
-            if (!browser.mobile) {
-                var $loader = $('<div class="meta-map-loader"><div class="valign"><div class="valign__middle">Click to use the map</div></div></div>');
+            if( !browser.mobile && self.config.click_to_use )
+            {
+                var $prevent = $('<div class="map-prevent"><div class="valign"><div class="valign__middle">Click to use the map</div></div></div>');
 
                 $map.append($loader, false);
-                $loader.click(function () { $loader.hide() });
+                $prevent.click(function () { $prevent.fadeOut(300) });
             }
 
-            self.context.gmap = $map.gmap3(self.config.map).then(function (result) {
+            self.context.gmap = $map.gmap3(self.config.map);
 
+            self.context.gmap.then(function ()
+            {
                 self.context.map = this.get(0);
 
-                google.maps.event.addListener(self.context.map, 'zoom_changed', function () {
+                google.maps.event.addListener(self.context.map, 'zoom_changed', function() {
 
                     var zoomLevel = self.context.map.getZoom();
-                    self.context.$map.trigger('map.zoom', [
-                        self.context.map,
-                        zoomLevel
-                    ]);
+                    self.context.$map.trigger('map.zoom',[self.context.map, zoomLevel]);
                 });
+
+                google.maps.event.addListener(self.context.map, 'bounds_changed', function() {
+
+                    self.context.$map.trigger('map.bounds_changed', [self.context.map]);
+                    setTimeout(self.updateLabel, 50);
+                });
+
+                google.maps.event.addListener(self.context.map, 'idle', function() {
+
+                    self.context.$map.trigger('map.idle', [self.context.map]);
+                    setTimeout(self.updateLabel, 50);
+                });
+
 
                 self.context.init = true;
 
-                if (callback) {
+                if( callback )
                     callback(self);
-                }
             });
         };
 
 
-        /**
-         *
-         */
-        self.zoomIn = function () {
-
-            if (self.context.map.getZoom() < 21) {
-                self.context.map.setZoom(self.context.map.getZoom() + 1);
-            }
+        self.updateLabel = function()
+        {
+            $map.find("[style*='custom-label']").addClass('map-label').removeAttr('style');
         };
 
 
         /**
          *
          */
-        self.zoomOut = function () {
+        self.zoomIn = function() {
 
-            if (self.context.map.getZoom() > 3) {
-                self.context.map.setZoom(self.context.map.getZoom() - 1);
-            }
+            if(self.context.map.getZoom() < 21 )
+                self.context.map.setZoom( self.context.map.getZoom()+1 );
         };
 
 
-        self.clearMarkers = function () {
+        /**
+         *
+         */
+        self.zoomOut = function() {
 
-            if (self.context.cluster) {
+            if(self.context.map.getZoom() >  3)
+                self.context.map.setZoom( self.context.map.getZoom()-1 );
+        };
+
+
+
+        self.clearMarkers = function(){
+
+            if( self.context.cluster ){
 
                 var markers = self.context.cluster.markers();
-                $.each(markers, function (i, marker) {
+                $.each(markers, function(i, marker){
 
                     self.context.cluster.remove(marker);
                 });
 
                 self.context.cluster = false;
             }
-            else {
+            else{
 
-                $.each(self.context.markers, function (i, marker) {
+                $.each(self.context.markers, function(i, marker){
 
                     marker.setMap(null);
                     delete self.context.markers[i];
@@ -214,28 +242,28 @@
         };
 
 
-        self.clearOverlay = function () {
 
-            if (self.context.overlay) {
+        self.clearOverlay = function(){
+
+            if( self.context.overlay )
                 self.context.overlay.setMap(null);
-            }
 
-            self.context.overlay     = false;
-            self.context.has_overlay = false;
+            self.context.overlay = false;
         };
 
 
-        self.setPosition = function (address, zoom) {
 
-            if (typeof address == 'object') {
+        self.setPosition = function(address, zoom){
 
-                var latLng = new google.maps.LatLng(address.coords.latitude, address.coords.longitude);
+            if( typeof address == 'object' ){
+
+                var latLng = new google.maps.LatLng(address.lat, address.lng);
                 self.context.map.panTo(latLng);
                 self.context.map.setZoom(zoom);
             }
-            else {
+            else{
 
-                self.context.gmap.latlng({address: address})
+                self.context.gmap.latlng({address:address})
                     .then(function (result) {
 
                         var latLng = new google.maps.LatLng(result.lat(), result.lng());
@@ -246,18 +274,18 @@
         };
 
 
-        self.hightlightMarker = function (id, status) {
 
-            if (self.context.init && self.context.markers.length > id) {
-                self.context.markers[id].setIcon(status ? self.config.marker_hover : self.config.marker);
-            }
+        self.hightlightMarker = function(id, status){
+
+            if( self.context.init && self.context.markers.length > id )
+                self.context.markers[id].setIcon( status ? self.context.markers[id].icons.hover : self.context.markers[id].icons.normal);
         };
 
 
         /**
          *
          */
-        self.reset = function () {
+        self.reset = function(){
 
             self.clearMarkers();
         };
@@ -266,129 +294,139 @@
         /**
          *
          */
-        self.addMyLocation = function (data, zoom) {
+        self.addMyLocation = function( data, zoom ){
 
-            self.context.gmap.marker(data).then(function (marker) {
+            self.context.gmap.marker(data).then(function(marker){
 
                 self.context.markers.push(marker);
 
-                if (typeof zoom != 'undefined') {
-                    self.setPosition({
-                        coords: {
-                            latitude : marker.position.lat(),
-                            longitude: marker.position.lng()
-                        }
-                    }, zoom);
-                }
+                if( typeof zoom != 'undefined')
+                    self.setPosition({lat:marker.position.lat(),lng:marker.position.lng()}, zoom);
             });
         };
 
 
-        self.addMarkers = function (markers, fit, icon) {
+
+        self.addMarkers = function( markers, fit, icon ){
 
             var labels = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
 
-            $.each(markers, function (i, marker) {
+            $.each(markers, function(i, marker)
+            {
+                marker.has_overlay = false;
+                marker.icon  = $.extend({}, self.context.marker);
 
-                marker.icon = $.extend({}, self.config.marker, icon ? icon(marker) : {});
+                marker.icons = {
+                    normal : $.extend({}, marker.icon),
+                    hover  : $.extend({}, marker.icon)
+                };
 
-                marker.icon_out   = $.extend({}, marker.icon);
-                marker.icon_hover = $.extend({}, marker.icon);
+                if( icon )
+                {
+                    var icon_path = icon(marker);
 
-                marker.icon_hover.url = marker.icon.hover_url;
+                    marker.icons.normal.url = icon_path.url;
+                    marker.icons.hover.url = icon_path.hover_url;
+                    marker.icon.url = marker.icons.normal.url;
+                }
+                else
+                {
+                    marker.icons.hover.url = self.config.marker.hover_url;
+                }
 
-                if (self.config.marker.label) {
-                    marker.label = {
-                        text    : labels[i++ % labels.length],
-                        color   : 'white',
-                        fontSize: '12px'
-                    };
+                if( self.config.marker.label )
+                    marker.label = {text: labels[i++ % labels.length], color: 'white', fontSize:'12px'};
+
+                if( 'position' in marker && 'address' in marker )
+                {
+                    marker._address = marker.address;
+                    delete marker.address;
                 }
             });
 
-            if ('cluster' in self.config) {
 
+            if( 'cluster' in self.config && self.config.cluster )
+            {
                 self.context.gmap.cluster({
-                    size   : self.config.cluster.size,
+                    size: self.config.cluster.size,
                     markers: markers,
-                    cb     : function (markers) {
+                    cb: function (markers) {
                         if (markers.length > 1) {
                             return {
-                                content: '<div class="' + self.config.cluster.class + '"><span>' + markers.length + '</span></div>',
-                                x      : self.config.cluster.x,
-                                y      : self.config.cluster.y
+                                content: '<div class="'+self.config.cluster.classname+'"><span>' + markers.length + '</span></div>',
+                                x: self.config.cluster.x,
+                                y: self.config.cluster.y
                             };
                         }
                     }
 
-                }).then(function (cluster) {
+                }).then(function(cluster){
 
                     self.context.markers = [];
                     self.context.cluster = cluster;
+                    setTimeout(self.updateLabel, 50);
+
+                }).on({
+                    click: function(marker, clusterOverlay, cluster, event){
+
+                        if (clusterOverlay)
+                            clusterOverlay.overlay.getMap().fitBounds(clusterOverlay.overlay.getBounds());
+                    }
                 });
             }
-            else {
-
-                self.context.gmap.marker(markers).then(function (markers) {
-
+            else
+            {
+                self.context.gmap.marker(markers).then(function(markers)
+                {
                     self.context.markers = self.context.markers.concat(markers);
-
+                    setTimeout(self.updateLabel, 50);
                 });
             }
 
             self.context.gmap.on({
 
-                mouseover: function (marker) {
+                mouseover: function(marker){
 
-                    if (typeof marker == 'undefined') {
+                    if( typeof marker == 'undefined')
                         return;
-                    }
 
-                    marker.setIcon(marker.icon_hover);
-                    self.context.$map.trigger('map.over', [
-                        self.context.map,
-                        marker.index
-                    ]);
+                    marker.setIcon(marker.icons.hover);
+                    self.context.$map.trigger('marker.over', [self.context.map, marker.index]);
                 },
-                mouseout : function (marker) {
+                mouseout: function(marker){
 
-                    if (typeof marker == 'undefined') {
+                    if( typeof marker == 'undefined')
                         return;
-                    }
 
-                    if (!self.context.has_overlay) {
-
-                        marker.setIcon(marker.icon_out);
-                        self.context.$map.trigger('map.out', [
-                            self.context.map,
-                            marker.index
-                        ]);
+                    if( !marker.has_overlay )
+                    {
+                        marker.setIcon(marker.icons.normal);
+                        self.context.$map.trigger('marker.out', [self.context.map, marker.index]);
                     }
                 },
-                click    : function (marker) {
+                click: function(marker){
 
-                    if (typeof marker == 'undefined') {
+                    if( typeof marker == 'undefined')
                         return;
-                    }
 
-                    if (self.context.marker) {
-                        self.context.marker.setIcon(self.context.marker.icon_out);
-                    }
+                    if( self.context.selected_marker )
+                        self.context.selected_marker.setIcon(self.context.selected_marker.icons.normal);
 
-                    self.context.marker = marker;
+                    self.context.selected_marker = marker;
 
-                    marker.setIcon(marker.icon_hover);
+                    marker.setIcon(marker.icons.hover);
 
-                    self.context.$map.trigger('map.click', [
-                        self.context.map,
-                        marker.index
-                    ]);
+                    self.context.$map.trigger('marker.click', [self.context.map, marker]);
 
-                    if (!browser.phone || ('phone' in self.config.overlay && self.config.overlay.phone)) {
+                    var enable = !browser.phone || ('phone' in self.config.overlay && self.config.overlay.phone);
+                    enable = enable && (!browser.mobile || ('mobile' in self.config.overlay && self.config.overlay.mobile));
+                    enable = enable && (!browser.tablet || ('tablet' in self.config.overlay && self.config.overlay.tablet));
 
+                    if( enable )
+                    {
                         self.clearOverlay();
 
-                        self.context.has_overlay = true;
+                        marker.has_overlay = true;
 
                         var html = self.config.overlay.html;
 
@@ -396,23 +434,24 @@
 
                         self.context.gmap.overlay({
                             position: marker.getPosition(),
-                            content : html,
-                            y       : self.config.overlay.y,
-                            x       : self.config.overlay.x
+                            content: html,
+                            y: self.config.overlay.y,
+                            x: self.config.overlay.x
 
-                        }).then(function (overlay) {
+                        }).then(function(overlay) {
 
                             self.context.overlay = overlay;
-                            overlay.$.find('[data-close]').click(function () {
 
-                                marker.setIcon(marker.icon_out);
-                                self.context.marker = false;
+                            self.context.$map.trigger('overlay.show', [self.context.map, marker]);
+
+                            overlay.$.find('[data-close]').click(function()
+                            {
+                                marker.setIcon(marker.icons.normal);
+                                self.context.selected_marker = false;
 
                                 self.clearOverlay();
-                                self.context.$map.trigger('map.out', [
-                                    self.context.map,
-                                    marker.index
-                                ]);
+                                marker.has_overlay = false;
+                                self.context.$map.trigger('overlay.hide', [self.context.map, marker]);
                             });
                         });
                     }
@@ -421,18 +460,16 @@
 
             /**/
 
-            if (fit) {
+            if( fit )
                 self.context.gmap.fit();
-            }
         };
 
-        self.__construct($map, template, config, callback);
+        self.__construct( $map, template, config, callback );
     };
 
     // disable gmap3 lib autoload
-    if (typeof $.gmap3 != 'undefined') {
+    if( typeof $.gmap3 != 'undefined' )
         $.gmap3(false);
-    }
 
     rocket = typeof rocket == 'undefined' ? {} : rocket;
     rocket.map = Map;