+++ src/js/metabolism/parallax.js
@@ -147,9 +147,9 @@
          */
         self.__construct = function () {
 
-            if ((browser && browser.mobile && !self.config.mobile) || typeof requestAnimationFrame == 'undefined') {
+
+            if( (browser && browser.mobile && !self.config.mobile) || typeof requestAnimationFrame == 'undefined' || $('html').hasClass('ie') )
                 return;
-            }
 
             self._setupEvents();