+++ src/js/metabolism/sliders.js
@@ -247,7 +247,6 @@
 
             self.context.$window
                 .resize(self._computeOffset)
-                .scroll(self._preload)
                 .scroll(self._checkVisibility);
 
             $(document).on('loaded', self._computeOffset);
@@ -296,6 +295,9 @@
             }
 
             self.context.is_visible = is_visible;
+
+            self._preload();
+            self._autoplay();
         };
 
 
@@ -405,13 +407,7 @@
                 self._addMod(self.context.$next_slide, 'slide', 'next');
             }
 
-            if (window.jQuery.fn.fit) {
-                self.context.$current_slide.find('[data-object_fit]').fit();
-
-                if (self.context.$next_slide) {
-                    self.context.$next_slide.find('[data-object_fit]').fit();
-                }
-            }
+            self.context.$window.resize();
 
             self._animate(animate, function () {
                 if (animate) {