+++ doc/changelog.md
@@ -4,13 +4,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 3.4.0 - 2013-03-14
+## 3.4.0 - 2013-03-20
 ### Added
 - Branch aliases for composer, instead of using `dev-dev` tag in composer, you can now use `3.4.0@dev`
 ### Changed
 - Composer Commands
   - `composer create` and `composer build` are now added to this project and can be used anywhere if installing from composer.
   - Symlinks creation is also moved into this project from boilerplate.
+- Hide-On behaviour
+- Media queries ( mobile = phone, tablet = small, ... )
+
+## 3.3.9 - 2017-03-17 ##
+### Fixed
+- IE Fit bug
+- no more autoplay while scrolling on slider
+### Removed
+- parallax on IE
+
+## 3.3.8 - 2017-03-15 ##
+### Fixed
+- has scrolled removed on top
+### Removed
+- hide on mobile, use hide on phone
+
+## 3.3.7 - 2017-03-15 ##
+### Fixed
+- Map marker hover error
+### Changed
+- map events
 
 ## 3.3.6 - 2017-03-10 ##
 ### Fixed